requirements.txt
python-dotenv
Flask
Flask-Cors
waitress
jsonpickle